Reorganised viewmodel
